requirements.txt
appdirs==1.4.4
asgiref==3.3.4
black==21.5b2
click==8.0.1
colorama==0.4.4
dj-database-url==0.5.0
Django==3.1.5
djangorestframework==3.12.4
gunicorn==20.1.0
mypy-extensions==0.4.3
pathspec==0.8.1
psycopg2-binary==2.8.6
pytz==2021.1
regex==2021.4.4
sqlparse==0.4.1
toml==0.10.2
whitenoise==5.2.0
